chore: use query renders for all my collection tabs
