+++ src/app/Scenes/MyProfile/MyProfileHeader.tsx
@@ -6,16 +6,20 @@ import {
   InstitutionIcon,
   MapPinIcon,
   SettingsIcon,
+  Spacer,
   Text,
   Touchable,
   useColor,
 } from "@artsy/palette-mobile"
+import { MyProfileHeaderQuery } from "__generated__/MyProfileHeaderQuery.graphql"
 import { MyProfileHeader_me$key } from "__generated__/MyProfileHeader_me.graphql"
 import { navigate } from "app/system/navigation/navigate"
 import { useLocalImageStorage } from "app/utils/LocalImageStore"
+import { withSuspense } from "app/utils/hooks/withSuspense"
+import { PlaceholderBox, PlaceholderText } from "app/utils/placeholders"
 import { useRefetch } from "app/utils/relayHelpers"
 import { Image, TouchableOpacity } from "react-native"
-import { useFragment } from "react-relay"
+import { useFragment, useLazyLoadQuery } from "react-relay"
 import { graphql } from "relay-runtime"
 import { normalizeMyProfileBio } from "./utils"
 
@@ -143,3 +147,36 @@ const myProfileHeaderFragment = graphql`
     createdAt
   }
 `
+
+const MyProfileHeaderPlaceholder: React.FC<{}> = () => {
+  return (
+    <Flex flex={1} px={2}>
+      <Spacer y={2} />
+      {/* icon, name, time joined */}
+      <Flex flexDirection="row">
+        <PlaceholderBox width={50} height={50} borderRadius={50} />
+        <Flex flex={1} justifyContent="center" ml={2}>
+          <PlaceholderText width={80} height={25} />
+          <PlaceholderText width={100} height={15} />
+        </Flex>
+        {/* settings icon */}
+        <PlaceholderBox width={20} height={20} />
+      </Flex>
+      <Spacer y={1} />
+    </Flex>
+  )
+}
+
+const myProfileHeaderQuery = graphql`
+  query MyProfileHeaderQuery {
+    me {
+      ...MyProfileHeader_me
+    }
+  }
+`
+
+export const MyProfileHeaderQueryRenderer = withSuspense(() => {
+  const data = useLazyLoadQuery<MyProfileHeaderQuery>(myProfileHeaderQuery, {})
+
+  return <MyProfileHeader me={data.me!} />
+}, MyProfileHeaderPlaceholder)
