+++ src/app/utils/hooks/withSuspense.tsx
@@ -1,0 +1,19 @@
+import { Spinner } from "@artsy/palette-mobile"
+import { ProvidePlaceholderContext } from "app/utils/placeholders"
+import { Suspense } from "react"
+
+export const withSuspense =
+  (Component: React.FC<any>, Fallback: React.FC<any> = () => <Spinner />) =>
+  (props: any) => {
+    return (
+      <Suspense
+        fallback={
+          <ProvidePlaceholderContext>
+            <Fallback {...props} />
+          </ProvidePlaceholderContext>
+        }
+      >
+        <Component {...props} />
+      </Suspense>
+    )
+  }
